# Use StrongBox<Rect> over object field in GlyphRun